new type hierarchy for features, not done fixing things so tests will fail. Still to do, at a minimum: fix tests, plus decide on terminology (encoder probably?)
