+++ src/features/bondfeature.jl
@@ -1,0 +1,6 @@
+
+struct BondFeature <: AbstractPairFeature
+    # stuff
+end
+
+# more stuff
