+++ src/features/ofm.jl
@@ -1,0 +1,6 @@
+
+struct OrbitalFieldMatrix <: AbstractEnvironmentFeature
+    # stuff
+end
+
+# more stuff
